+++ src/main/kotlin/cloud/glitchdev/rfu/achievement/interfaces/IAchievement.kt
@@ -14,6 +14,6 @@ interface IAchievement {
     val isCompleted: Boolean
     val isCheated: Boolean
     val progress: Float // 0.0 to 1.0
-    val currentProgress: Int get() = if (isCompleted) 1 else 0
-    val targetProgress: Int get() = 1
+    val currentProgress: Long get() = if (isCompleted) 1L else 0L
+    val targetProgress: Long get() = 1L
 }
