+++ src/main/kotlin/cloud/glitchdev/rfu/achievement/types/NumericStageAchievement.kt
@@ -3,18 +3,18 @@ package cloud.glitchdev.rfu.achievement.types
 import cloud.glitchdev.rfu.achievement.AchievementProvider
 
 abstract class NumericStageAchievement : StageAchievement() {
-    abstract fun getTargetCountForStage(stage: Int): Int
+    abstract fun getTargetCountForStage(stage: Int): Long
 
     open val resetCountOnStageAdvance: Boolean = true
 
     override fun onStageChanged(oldStage: Int, newStage: Int) {
         if (resetCountOnStageAdvance) {
-            currentCount = 0
+            currentCount = 0L
         }
     }
 
-    open val targetCount: Int get() = getTargetCountForStage(currentStage)
-    var currentCount: Int = 0
+    open val targetCount: Long get() = getTargetCountForStage(currentStage)
+    var currentCount: Long = 0L
         protected set(value) {
             field = value
             if (field >= targetCount) {
@@ -24,18 +24,18 @@ abstract class NumericStageAchievement : StageAchievement() {
             }
         }
 
-    override val targetProgress: Int get() = targetCount
-    override val currentProgress: Int get() = currentCount
+    override val targetProgress: Long get() = targetCount
+    override val currentProgress: Long get() = currentCount
 
     override var _progress: Float = 0.0f
-        get() = if (targetCount > 0) currentCount.toFloat() / targetCount.toFloat() else 0.0f
+        get() = if (targetCount > 0L) currentCount.toFloat() / targetCount.toFloat() else 0.0f
 
-    fun addProgress(amount: Int = 1) {
+    fun addProgress(amount: Long = 1L) {
         if (isCompleted) return
         currentCount += amount
     }
 
-    fun debugAddProgress(amount: Int = 1) {
+    fun debugAddProgress(amount: Long = 1L) {
         if (isCompleted) return
         markAsCheated()
         addProgress(amount)
@@ -47,14 +47,14 @@ abstract class NumericStageAchievement : StageAchievement() {
     }
 
     override fun debugReset() {
-        currentCount = 0
+        currentCount = 0L
         currentStage = 1
         super.debugReset()
     }
 
     override fun loadState(progressData: Map<String, Any>) {
         super.loadState(progressData)
-        currentCount = (progressData["currentCount"] as? Number)?.toInt() ?: 0
+        currentCount = (progressData["currentCount"] as? Number)?.toLong() ?: 0L
     }
 
     override fun saveState(): Map<String, Any> {
